moved validate_webhook and authenticate out in to a client of their own
